+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 For the moment we use git hashes as version, since no one complained about it 👍
 
+## Measure tools and better continuous integration (xxxxx XX/04/2022)
+
+- Users can now measure existing shapes. Length for lines, Areas for polygons.
+- Add "light" pipelines of continuous integration, per default. These pipelines execute: lint, build, unit and 
+integration tests, but not end to end tests or performance tests. It allows executing CI even on Gitlab  
+shared runners.
+- Some entities are not available globally: FeatureWrapper, LayerWrapper, MapWrapper, ... for hacking in consoles.
+
 ## Text frames style (606a2fd6 03/04/2022)
 
 - Style options for text frames. You can now open a modal on text frames to change background color,